Bug fixing

+++ cloudmesh/key/command/key.py
@@ -33,18 +33,21 @@ class KeyCommand(PluginCommand):
         m = Manager()
 
         #TODO - Make sure they set the master parameter
-        if arguments.gatherkeys:
+        if arguments.gatherkeys and master != "":
             os.system("rm -rf /home/pi/keys")
             os.system("mkdir /home/pi/keys")
             os.system("rm /home/pi/.ssh/authorized_keys")
             for i in ips:
                 ip = i
 
-                #Generate key for worker
+                #When burning workers using cm-pi-burn, it makes .ssh owned by root instead of pi
+                #This fixes permissions if it needs to be fixed
                 command = "ssh " + ip + " sudo chown pi .ssh"
                 os.system(command)
                 command = "ssh " + ip + " sudo chown pi .ssh/authorized_keys"
                 os.system(command)
+
+                #Generate key for worker
                 command = "ssh " + ip + " ssh-keygen"
                 os.system(command) 
 
@@ -54,25 +57,27 @@ class KeyCommand(PluginCommand):
                 command = "cat /home/pi/keys/" + ip + "pk >> /home/pi/.ssh/authorized_keys"
                 os.system(command)
 
-            os.system("cp /home/pi/.ssh/authorized_keys /home/pi/keys/")
+            #Once all public keys gathered, this is the redistribution step back to workers
             for i in ips:
                 ip = i
-                #Redistribute all public keys (including master) back to worker's authorized keys
-                command = "scp /home/pi/keys/authorized_keys pi@" + ip + ":/home/pi/public_keys"
+                #Copy over authorized keys file from master that includes all worker's public keys
+                command = "scp /home/pi/.ssh/authorized_keys pi@" + ip + ":/home/pi/public_keys"
                 os.system(command)
 
-                #Copy over master public key
+                #Copy over master public key to worker
                 command = "scp /home/pi/.ssh/id_rsa.pub pi@" + ip + ":/home/pi/master_pk"
                 os.system(command)
+
+                #Copy master public key into file with all other worker  public keys
                 command = "ssh " + ip + " 'cat /home/pi/master_pk >> /home/pi/public_keys'"
                 os.system(command) 
+
+                #Write over worker's authorized_keys with file containing all worker's and master's public key
                 command = "ssh " + ip + " cp /home/pi/public_keys /home/pi/.ssh/authorized_keys"
                 os.system(command)
 
-                #Delete public_keys & master_pk files that was transfered over after contents of it is put in authorized_keys
+                #Delete public_keys & master_pk tmp files
                 os.system("ssh " + ip + " rm /home/pi/public_keys")
                 os.system("ssh " + ip + " rm /home/pi/master_pk")
-
-           #TODO - Do I need to delete the authorized_keys file I saved to /home/pi/keys on master?
 
         return ""
